Disambiguation of Lab 8
As it stands, I find the challenge part of lab 8 to be ambiguous: "possibly uninitialized" is the same as "may be uninitialized", so it could be interpreted as the challenge being a mere switch from error reporting to warning reporting (which is trivial). However, I think that you mean to say that warnings should only be raised for "maybe" cases and not "definite" cases, while errors should be changed to only respond to "definite" cases. Am I interpreting this correctly?